backend/requirements.txt
cmake
screen
django
django-admin
djangorestframework
numpy 
pandas 
scikit-learn